requirements.txt
feedparser
Flask
Pillow
pystray
psutil
yfinance
pytz
requests
beautifulsoup4